Typecast return value
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxPython/Phoenix/trunk@71294 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ etg/bitmap.py
@@ -62,7 +62,7 @@ def run():
         doc='MSW-only method to fetch the windows handle for the bitmap.',
         body="""\
             #ifdef __WXMSW__
-                return self->GetHandle();
+                return (long)self->GetHandle();
             #else
                 return 0;
             #endif
